+++ src/core/wrapper.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import store from '../redux/store';
 
-const AppProvider = ({children}) => <Provider store={store}>{children}</Provider>;
+import configureStore from '../redux/configureStore';
+
+const AppProvider = ({children}) => <Provider store={configureStore()}>{children}</Provider>;
 export default AppProvider;